NXDRIVE-1090: Move out testing methods out of Remote to RemoteTest
